+++ e2e/_archive/dev-pages/repo_dev_page.ts
@@ -1,0 +1,10 @@
+const { expect } = require('@playwright/test');
+
+export default class RepoDevPage {
+    private _page: any;
+
+    constructor(page, route) {
+        this._page = page;
+    }
+
+}
